Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,9 +2,9 @@
 
 
 int main() {
-    int a,s,d;
+    float a,s,d;
     float avg;
-    scanf("%d %d %d",&a,&s,&d);
+    scanf("%f %f %f",&a,&s,&d);
     avg=(a+s+d)/3;
     printf("Average: %.2f",avg);
     return 0;
